Port r730523 from tcnative trunk to tc trunk. Add Thread.java and Buffer.java. Change SSL.randSet() javadoc to the tcnative one. Add Socket.sendib() and Socket.sendibb from tcnative.
All renames methods and method signature changes
are compatible, because those methods are not used
by Tomcat.

Changes of return types are compatible, because
the return values have not been used.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@730544 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/tomcat/jni/Registry.java
@@ -156,7 +156,7 @@ public class Registry {
      * @param val The the value to set
      * @return If the function succeeds, the return value is 0
      */
-    public static native int setValueJ(long key, String name, int val);
+    public static native int setValueJ(long key, String name, long val);
 
     /**
      * Set the Registry value for REG_SZ
